Remove --update-cache hardcoded option
If apk is run with the default options, there will not be any cache in
the root's /var/cache/apk so the cache will be downloaded anyway.

If a cache is specified through APK_OPTS then it might not be expected
that the index be refreshed everytime, especially if other extra options
like --cache-max-age are set.

If --update-cache had been specified before APK_OPTS then it could be
overriden but in its current form it cannot be, and it's not useful
anywhere, so remove it.